requirements.txt
numpy
scipy
attrs
mpmath
